Move wait_for_shared_memory to standalone file
